made changes

+++ src/main/java/com/maveric/userservice/service/UserServiceImpl.java
@@ -1,6 +1,8 @@
 package com.maveric.userservice.service;
 import com.maveric.userservice.dto.UserDto;
 import com.maveric.userservice.exception.ResourceNotFoundException;
+import com.maveric.userservice.exception.UserAlreadyExistException;
+import com.maveric.userservice.exception.UserNotFoundException;
 import com.maveric.userservice.model.User;
 import com.maveric.userservice.repository.UserRepository;
 import lombok.AllArgsConstructor;
@@ -11,6 +13,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -47,38 +50,74 @@ public class UserServiceImpl implements UserService{
 
     @Override
     public UserDto getUserDetails(String id) {
-       return userRepository.findById(Long.valueOf(id)).map(this::convertEntityToDto).get();
+
+        User userResult = userRepository.findById(Long.valueOf(id)).orElseThrow(() -> new UserNotFoundException("User not found with id " + id));
+        return userRepository.findById(Long.valueOf(id)).map(this::convertEntityToDto).get();
     }
 
     @Override
     public UserDto createUser(UserDto userDto) {
-        User user= convertDtoToEntity(userDto);
-        User newuser= userRepository.save(user);
-        UserDto userresponce= convertEntityToDto(newuser);
-        return userresponce;
+
+        User userResult = userRepository.findByEmail(userDto.getEmail());
+        if (userResult == null) {
+                    User user= convertDtoToEntity(userDto);
+                    User newuser= userRepository.save(user);
+                     UserDto userresponce= convertEntityToDto(newuser);
+                    return userresponce;
+        } else {
+          //  log.error("User Already Exist for this emailId");
+            throw new UserAlreadyExistException("User Already Exist! for this emailId");
+        }
+
+//        User user= convertDtoToEntity(userDto);
+//        User newuser= userRepository.save(user);
+//        UserDto userresponce= convertEntityToDto(newuser);
+//        return userresponce;
     }
+
+
     @Override
     public UserDto getUserByEmail(String email) {
-          User user = userRepository.findByEmail(email);
-          return convertEntityToDto(user);
+        User userResult = userRepository.findByEmail(email);
+        if (userResult != null)
+        {
+            return convertEntityToDto(userResult);
+        }
+        else
+        {
+            throw new UserNotFoundException("User not found with id " + email);
+        }
+
+
+
+        //  User user = userRepository.findByEmail(email);
+        //  return convertEntityToDto(user);
     }
     @Override
     public String deleteUser(String id) {
-                userRepository.deleteById(Long.valueOf(id));
+
+        if (userRepository.findById(Long.valueOf(id)).isEmpty()) {
+            throw new UserNotFoundException("User not found with id " + id);
+        }
+        userRepository.deleteById(Long.valueOf(id));
         return "User deleted successfully.";
+//
+//                userRepository.deleteById(Long.valueOf(id));
+//        return "User deleted successfully.";
     }
 
     @Override
     public UserDto updateUser(String id, UserDto userDto) {
         User updateUser = userRepository.findById(Long.valueOf(id))
-                .orElseThrow(() -> new ResourceNotFoundException("User not exit with id" +id));
+                .orElseThrow(() -> new UserNotFoundException("User not exit with id" + id));
         updateUser.setFirstName(userDto.getFirstName());
         updateUser.setLastName(userDto.getLastName());
         updateUser.setMiddleName(userDto.getMiddleName());
         updateUser.setPhoneNumber(userDto.getPhoneNumber());
         updateUser.setEmail(userDto.getEmail());
         updateUser.setAddress(userDto.getAddress());
-        updateUser.setDateOfBirth(userDto.getDateOfBirth());
+       updateUser.setDateOfBirth(userDto.getDateOfBirth());
+       // updateUser.setDateOfBirth(userDto.getDateOfBirth());
         updateUser.setGender(userDto.getGender());
         updateUser.setId(userDto.getId());
         userRepository.save(updateUser);
@@ -96,6 +135,7 @@ public class UserServiceImpl implements UserService{
         userDto.setEmail(user.getEmail());
         userDto.setAddress(user.getAddress());
         userDto.setDateOfBirth(user.getDateOfBirth());
+      //  userDto.setDateOfBirth(user.getDateOfBirth());
         userDto.setGender(user.getGender());
         userDto.setId(user.getId());
         userDto.setPassword(user.getPassword());
@@ -110,7 +150,8 @@ public class UserServiceImpl implements UserService{
         user.setEmail(userDto.getEmail());
         user.setAddress(userDto.getAddress());
         user.setDateOfBirth(userDto.getDateOfBirth());
-        user.setGender(userDto.getGender());
+         // user.setDateOfBirth(userDto.getDateOfBirth());
+          user.setGender(userDto.getGender());
         user.setId(userDto.getId());
         user.setPassword(userDto.getPassword());
         user.setPassword(new BCryptPasswordEncoder().encode(userDto.getPassword()));
